wip: test preload files
- wip commit for testing preload files with electron-mocha. I think the test build is correct, but it wont run with regular mocha because the node version differs beween eletron and local node. I could rebuild for test vs dev but would be slow and maybe fragile constantly going back and forth. One alternative is to use electrons version of node; you can do this but need to then call mocha and do some electron setup; this is what elecgtron-mocha does. But as of this commit its not working -- it doesnt find the test files for one reason or another (outputs 0 tests ran).
