+++ src/views/edit/index.test.ts
@@ -1,9 +1,11 @@
-import { test } from "mocha";
+import { test, suite } from "mocha";
 
-test("loads");
-test("has loading state");
-test("saves document");
-test("cannot edit while saving");
-test("saving a document twice updates it...");
-test("saving error is surfaced");
-test("loading error is surfaced");
+suite("views.edit", function () {
+  test("loads");
+  test("has loading state");
+  test("saves document");
+  test("cannot edit while saving");
+  test("saving a document twice updates it...");
+  test("saving error is surfaced");
+  test("loading error is surfaced");
+});
